commit 29.12.2020

+++ src/main/java/com/nesterov/university/service/AudienceService.java
@@ -6,10 +6,8 @@ import java.util.List;
 import java.util.Optional;
 import org.springframework.stereotype.Component;
 import com.nesterov.university.dao.AudienceDao;
-import com.nesterov.university.dao.exceptions.NotCreateException;
-import com.nesterov.university.dao.exceptions.NotDeleteException;
 import com.nesterov.university.dao.exceptions.NotFoundEntitiesException;
-import com.nesterov.university.dao.exceptions.NotPresentEntityException;
+import com.nesterov.university.dao.exceptions.NotFoundException;
 import com.nesterov.university.dao.exceptions.NotUniqueRoomNumberException;
 import com.nesterov.university.model.Audience;
 
@@ -22,28 +20,32 @@ public class AudienceService {
 		this.audienceDao = audienceDao;
 	}
 
-	public void create(Audience audience) throws NotUniqueRoomNumberException, NotCreateException {
-		if (isUniqueRoomNumber(audience)) {
+	public void create(Audience audience) {
+		if (isUniqueRoomNumber(audience).getId() == audience.getId()) {
 			audienceDao.create(audience);
 		}
 	}
 
-	public Audience get(long id) throws NotPresentEntityException {
+	public Audience get(long id) {
 		String message = format("Audience with id = '%s' not found", id);
-		return audienceDao.get(id).orElseThrow(() -> new NotPresentEntityException(message));
+		return audienceDao.get(id).orElseThrow(() -> new NotFoundException(message));
 	}
 
-	public void delete(long id) throws NotDeleteException {
+	public void delete(long id) {
+		if (!audienceDao.get(id).isPresent()) {
+			String message = format("Audience with id = '%s' not found", id);
+			throw new NotFoundException(message);
+		}
 		audienceDao.delete(id);
 	}
 
-	public void update(Audience audience) throws NotUniqueRoomNumberException {
-		if (isUniqueRoomNumber(audience)) {
+	public void update(Audience audience) {
+		if (isUniqueRoomNumber(audience).getId() == audience.getId()) {
 			audienceDao.update(audience);
 		}
 	}
 
-	public List<Audience> getAll() throws NotFoundEntitiesException {
+	public List<Audience> getAll() {
 		List<Audience> audiences = audienceDao.findAll();
 		if (audiences.isEmpty()) {
 			throw new NotFoundEntitiesException("Not found audiences");
@@ -51,12 +53,12 @@ public class AudienceService {
 		return audiences;
 	}
 
-	private boolean isUniqueRoomNumber(Audience audience) throws NotUniqueRoomNumberException {
+	private Audience isUniqueRoomNumber(Audience audience) {
 		Optional<Audience> optionalAudience = audienceDao.findByRoomNumber(audience.getRoomNumber());
-		if (optionalAudience.isPresent() && optionalAudience.orElse(new Audience()).getId() != audience.getId()) {
+		if (!optionalAudience.isPresent()) {
 			String message = format("Not unique roomNumber = '%s'", audience.getRoomNumber());
 			throw new NotUniqueRoomNumberException(message);
 		}
-		return !optionalAudience.isPresent() || optionalAudience.orElse(new Audience()).getId() == audience.getId();
+		return optionalAudience.get();
 	}
 }
