+++ src/main/java/com/nesterov/university/service/GroupService.java
@@ -6,9 +6,8 @@ import java.util.List;
 import java.util.Optional;
 import org.springframework.stereotype.Component;
 import com.nesterov.university.dao.GroupDao;
-import com.nesterov.university.dao.exceptions.NotCreateException;
-import com.nesterov.university.dao.exceptions.NotDeleteException;
 import com.nesterov.university.dao.exceptions.NotFoundEntitiesException;
+import com.nesterov.university.dao.exceptions.NotFoundException;
 import com.nesterov.university.dao.exceptions.NotPresentEntityException;
 import com.nesterov.university.dao.exceptions.NotUniqueNameException;
 import com.nesterov.university.model.Group;
@@ -22,28 +21,32 @@ public class GroupService {
 		this.groupDao = groupDao;
 	}
 
-	public void create(Group group) throws NotCreateException, NotUniqueNameException {
-		if (isUniqueName(group)) {
+	public void create(Group group) {
+		if (isUniqueName(group).getId() == group.getId()) {
 			groupDao.create(group);
 		}
 	}
 
-	public void delete(long id) throws NotDeleteException {
+	public void delete(long id) {
+		if (!groupDao.get(id).isPresent()) {
+			String message = format("Group with id = '%s' not found", id);
+			throw new NotFoundException(message);
+		}
 		groupDao.delete(id);
 	}
 
-	public Group get(long id) throws NotPresentEntityException {
+	public Group get(long id) {
 		String message = format("Group with id = '%s' not found", id);
 		return groupDao.get(id).orElseThrow(() -> new NotPresentEntityException(message));
 	}
 
-	public void update(Group group) throws NotUniqueNameException {
-		if (isUniqueName(group)) {
+	public void update(Group group) {
+		if (isUniqueName(group).getId() == group.getId()) {
 			groupDao.update(group);
 		}
 	}
 
-	public List<Group> getAll() throws NotFoundEntitiesException {
+	public List<Group> getAll() {
 		List<Group> groups = groupDao.findAll();
 		if (groups.isEmpty()) {
 			throw new NotFoundEntitiesException("Not Found groups");
@@ -51,12 +54,12 @@ public class GroupService {
 		return groups;
 	}
 
-	private boolean isUniqueName(Group group) throws NotUniqueNameException {
+	private Group isUniqueName(Group group) {
 		Optional<Group> optionalGroup = groupDao.findByName(group.getName());
-		if (optionalGroup.isPresent() && optionalGroup.orElse(null).getId() != group.getId()) {
+		if (!optionalGroup.isPresent()) {
 			String message = String.format("Not unique groupName = '%s'", group.getName());
 			throw new NotUniqueNameException(message);
 		}
-		return !optionalGroup.isPresent() || optionalGroup.orElse(null).getId() == group.getId();
+		return optionalGroup.get();
 	}
 }
