+++ src/main/java/com/nesterov/university/service/LessonTimeService.java
@@ -5,10 +5,7 @@ import static java.lang.String.format;
 import java.util.List;
 import org.springframework.stereotype.Component;
 import com.nesterov.university.dao.LessonTimeDao;
-import com.nesterov.university.dao.exceptions.NotCreateException;
-import com.nesterov.university.dao.exceptions.NotDeleteException;
-import com.nesterov.university.dao.exceptions.NotFoundEntitiesException;
-import com.nesterov.university.dao.exceptions.NotPresentEntityException;
+import com.nesterov.university.dao.exceptions.NotFoundException;
 import com.nesterov.university.dao.exceptions.NotRightTimeException;
 import com.nesterov.university.model.LessonTime;
 
@@ -21,36 +18,40 @@ public class LessonTimeService {
 		this.lessonTimeDao = lessonTimeDao;
 	}
 
-	public void create(LessonTime lessonTime) throws NotRightTimeException, NotCreateException {
+	public void create(LessonTime lessonTime) {
 		if (isRightTime(lessonTime)) {
 			lessonTimeDao.create(lessonTime);
 		}
 	}
 
-	public void delete(long id) throws NotDeleteException {
+	public void delete(long id) {
+		if (!lessonTimeDao.get(id).isPresent()) {
+			String message = format("LessonTime with id = '%s' not found", id);
+			throw new NotFoundException(message);
+		}
 		lessonTimeDao.delete(id);
 	}
 
-	public LessonTime get(long id) throws NotPresentEntityException {
+	public LessonTime get(long id) {
 		String message = format("LessonTime with id = '%s' not found", id);
-		return lessonTimeDao.get(id).orElseThrow(() -> new NotPresentEntityException(message));
+		return lessonTimeDao.get(id).orElseThrow(() -> new NotFoundException(message));
 	}
 
-	public void update(LessonTime lessonTime) throws NotRightTimeException {
+	public void update(LessonTime lessonTime) {
 		if (isRightTime(lessonTime)) {
 			lessonTimeDao.update(lessonTime);
 		}
 	}
 
-	public List<LessonTime> getAll() throws NotFoundEntitiesException {
+	public List<LessonTime> getAll() {
 		List<LessonTime> lessonTimes = lessonTimeDao.findAll();
 		if (lessonTimes.isEmpty()) {
-			throw new NotFoundEntitiesException("Lessontimes not found");
+			throw new NotFoundException("Lessontimes not found");
 		}
 		return lessonTimes;
 	}
 
-	private boolean isRightTime(LessonTime lessonTime) throws NotRightTimeException {
+	private boolean isRightTime(LessonTime lessonTime) {
 		if (lessonTime.getStart().isAfter(lessonTime.getEnd())) {
 			throw new NotRightTimeException("This is not right time of lesson");
 		}
