+++ src/main/java/com/nesterov/university/service/SubjectService.java
@@ -6,10 +6,7 @@ import java.util.List;
 import java.util.Optional;
 import org.springframework.stereotype.Component;
 import com.nesterov.university.dao.SubjectDao;
-import com.nesterov.university.dao.exceptions.NotCreateException;
-import com.nesterov.university.dao.exceptions.NotDeleteException;
-import com.nesterov.university.dao.exceptions.NotFoundEntitiesException;
-import com.nesterov.university.dao.exceptions.NotPresentEntityException;
+import com.nesterov.university.dao.exceptions.NotFoundException;
 import com.nesterov.university.dao.exceptions.NotUniqueNameException;
 import com.nesterov.university.model.Subject;
 
@@ -22,49 +19,52 @@ public class SubjectService {
 		this.subjectDao = subjectDao;
 	}
 
-	public void create(Subject subject) throws NotCreateException, NotUniqueNameException {
-		if (isUniqueName(subject)) {
+	public void create(Subject subject) {
+		if (isUniqueName(subject).getId() == subject.getId()) {
 			subjectDao.create(subject);
 		}
 	}
 
-	public void delete(long id) throws NotDeleteException {
+	public void delete(long id) {
+		if (!subjectDao.get(id).isPresent()) {
+			throw new NotFoundException("Subject not found");
+		}
 		subjectDao.delete(id);
 	}
 
-	public Subject get(long id) throws NotPresentEntityException {
-		return subjectDao.get(id).orElseThrow(() -> new NotPresentEntityException("Subject not found"));
+	public Subject get(long id) {
+		return subjectDao.get(id).orElseThrow(() -> new NotFoundException("Subject not found"));
 	}
 
-	public void update(Subject subject) throws NotUniqueNameException {
-		if (isUniqueName(subject)) {
+	public void update(Subject subject) {
+		if (isUniqueName(subject).getId() == subject.getId()) {
 			subjectDao.update(subject);
 		}
 	}
 
-	public List<Subject> getAll() throws NotFoundEntitiesException {
+	public List<Subject> getAll() {
 		List<Subject> subjects = subjectDao.findAll();
 		if (subjects.isEmpty()) {
-			throw new NotFoundEntitiesException("Not found subjects");
+			throw new NotFoundException("Not found subjects");
 		}
 		return subjects;
 	}
 
-	public List<Subject> findByTeacherId(long id) throws NotFoundEntitiesException {
+	public List<Subject> findByTeacherId(long id) {
 		List<Subject> subjects = subjectDao.findByTeacherId(id);
 		if (subjects.isEmpty()) {
 			String message = format("Not found subjects by teacherId = '%s'", id);
-			throw new NotFoundEntitiesException(message);
+			throw new NotFoundException(message);
 		}
 		return subjects;
 	}
 
-	private boolean isUniqueName(Subject subject) throws NotUniqueNameException {
+	private Subject isUniqueName(Subject subject) {
 		Optional<Subject> founded = subjectDao.findByName(subject.getName());
-		if (founded.isPresent() && founded.orElse(new Subject()).getId() != subject.getId()) {
+		if (!founded.isPresent()) {
 			String message = format("groupName '%s' is not unique", subject.getName());
 			throw new NotUniqueNameException(message);
 		}
-		return !founded.isPresent() || founded.orElse(new Subject()).getId() == subject.getId();
+		return founded.get();
 	}
 }
